api: split DuckDBVector.ts into one file per vector class
Move each vector subclass into its own file under api/src/vectors/, matching
the api/src/values/ and api/src/types/ layout. Shared internals move to
sibling modules (DuckDBValidity, dataAccessors), and the abstract base lives
in vectors/DuckDBVector.ts.

The DuckDBVector.create() factory switch moves to vectors/createVector.ts and
is wired in through an internal vectorRegistry holder. This breaks what would
otherwise be a module-initialization cycle: the base statically importing its
subclasses (for the factory) while each subclass statically imports the base
(to extend it) throws at load time ("Cannot access 'DuckDBVector' before
initialization"), independent of when create() runs. The base now imports only
the registry; createVector imports the base + subclasses and assigns the impl.

api/src/DuckDBVector.ts becomes a re-export of ./vectors, so existing import
paths keep working unchanged. Public export set is identical (base + 42
subclasses); the registry/factory/accessor/validity modules are internal and
not re-exported from the barrel.

Pure structural change: no behavior or public API changes.
