+++ api/src/vectors/DuckDBDecimal16Vector.ts
@@ -1,0 +1,104 @@
+import duckdb from '@duckdb/node-bindings';
+import {
+  DuckDBDecimalType,
+} from '../DuckDBType';
+import {
+  DuckDBDecimalValue,
+} from '../values';
+import { DuckDBVector } from './DuckDBVector';
+import { DuckDBValidity } from './DuckDBValidity';
+import {
+  getDecimal16,
+  getInt16,
+  setInt16,
+  vectorData,
+} from './dataAccessors';
+
+export class DuckDBDecimal16Vector extends DuckDBVector<DuckDBDecimalValue> {
+  private readonly decimalType: DuckDBDecimalType;
+  private readonly dataView: DataView;
+  private readonly validity: DuckDBValidity;
+  private readonly vector: duckdb.Vector;
+  private readonly _itemCount: number;
+  constructor(
+    decimalType: DuckDBDecimalType,
+    dataView: DataView,
+    validity: DuckDBValidity,
+    vector: duckdb.Vector,
+    itemCount: number
+  ) {
+    super();
+    this.decimalType = decimalType;
+    this.dataView = dataView;
+    this.validity = validity;
+    this.vector = vector;
+    this._itemCount = itemCount;
+  }
+  static fromRawVector(
+    decimalType: DuckDBDecimalType,
+    vector: duckdb.Vector,
+    itemCount: number
+  ): DuckDBDecimal16Vector {
+    const data = vectorData(vector, itemCount * 2);
+    const dataView = new DataView(
+      data.buffer,
+      data.byteOffset,
+      data.byteLength
+    );
+    const validity = DuckDBValidity.fromVector(vector, itemCount);
+    return new DuckDBDecimal16Vector(
+      decimalType,
+      dataView,
+      validity,
+      vector,
+      itemCount
+    );
+  }
+  public override get type(): DuckDBDecimalType {
+    return this.decimalType;
+  }
+  public override get itemCount(): number {
+    return this._itemCount;
+  }
+  public override getItem(itemIndex: number): DuckDBDecimalValue | null {
+    return this.validity.itemValid(itemIndex)
+      ? getDecimal16(this.dataView, itemIndex * 2, this.decimalType)
+      : null;
+  }
+  public getScaledValue(itemIndex: number): number | null {
+    return this.validity.itemValid(itemIndex)
+      ? getInt16(this.dataView, itemIndex * 2)
+      : null;
+  }
+  public override setItem(itemIndex: number, value: DuckDBDecimalValue | null) {
+    if (value != null) {
+      setInt16(this.dataView, itemIndex * 2, Number(value.value));
+      this.validity.setItemValid(itemIndex, true);
+    } else {
+      this.validity.setItemValid(itemIndex, false);
+    }
+  }
+  public override flush() {
+    duckdb.copy_data_to_vector(
+      this.vector,
+      0,
+      this.dataView.buffer as ArrayBuffer,
+      this.dataView.byteOffset,
+      this.dataView.byteLength
+    );
+    this.validity.flush(this.vector);
+  }
+  public override slice(offset: number, length: number): DuckDBDecimal16Vector {
+    return new DuckDBDecimal16Vector(
+      this.decimalType,
+      new DataView(
+        this.dataView.buffer,
+        this.dataView.byteOffset + offset * 2,
+        length * 2
+      ),
+      this.validity.slice(offset, length),
+      this.vector,
+      length
+    );
+  }
+}
